Install nvim first and add fontsize adjustment instructions
Add instructions to install nvim very first. This allows me to edit text files with nvim immediately. Previously I used vim prior to nvim being installed, but the problem with this approach is some .vim files will be created that aren't bad, but are just clutter that we can avoid by adding nvim first.

Since this is the server branch add instructions on how to adjust terminal fontsize.